back-upc-fertismart/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-background-tasks==1.2.5
django-crontab==0.7.1
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-import-export==2.4.0
django-jet==1.0.8
djangorestframework==3.12.2
drf-access-policy==0.8.1
drf-yasg==1.20.0
et-xmlfile==1.0.1
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.2
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.8
Pillow==9.3.0
pymongo==3.11.0
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
rest-condition==1.0.3
six==1.13.0
sqlparse==0.4.1
shortuuid==1.0.11
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.2
xlrd==2.0.1
xlwt==1.3.0
django-shortuuidfield==0.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
pandas==1.3.5
djangorestframework-api-key==2.0.0
django-extensions==3.1.3
pdfkit==1.0.0
django-redis==4.12.1
redis==3.5.3
django-rq==2.5.1
scikit-learn==1.3.2
xhtml2pdf==0.2.16